The GitHub URL for files within the project was not as expected, nor did a corrected URL behave as expected.
The whole help mess started with Microsoft obsoleting the HTML help workshop.  This meant that the chmProcessor tool could no longer build the CHM help file for distribution with the program.  Now, to show help, Microsoft wants a URL to an HTML file somewhere.

Attempting to provide help by referring to the URL of the HTML help index file in the GitHub repository didn't work, though, because GitHub serves up a generated page with the HTML text of the file; it does not serve up the file itself.

The current resolution is to instruct the installer to install all of the HTML help files along with the program, then to refer to a browser to the local copy of the help files when the user asks to have the help shown.  This seems to work, but I suspect that if revising the documentation causes the help file generation process to generate files with different names than the ones that the installer is now installing that the installer itself will have to be modified to ensure that all the help files get installed, which could be ugly.