Changes to the ColorFromOverlayForm and its handling
Implemented the ability to use the L*, a*, and b* sliders and checkboxes to adjust the colors computed from the overlay prior to matching them to the palette colors.

Added a progress bar so that there is some indication that something is happening as the preview image is being computed.

Draw the preview image only once during form initialization, and delay the draw until the (blank) form is initially shown.  This provides a faster response to the menu choice having been selected, and also allows the user to see the progress bar move as the initial preview image is computed.