Improve efficiency of ColorFromOverlayForm preview image generation.
Improved the efficiency of preview image generation in the ColorFromOverlayForm by computing the average overlay color for each ring only once, when the form is first shown.  This operation was by far the longest-running component of preview image generation.

Refactored re-generation of the preview image (and also the final coloring of the design) to  apply the ring size filter, selected palette section, and color transformation to the already-computed average color information.